Rewrite HeroBackground from scratch using Three.js/WebGL, add playground (#30)
This PR includes a rewrite of the `<HeroBackground />` component that appears on the home page and 404 pages. The Canvas is a custom implementation built on [Three.js](https://threejs.org/)/WebGL, in the `<WaveCanvas>` component and `WaveRenderer` classes.

In reality, Three.js is very heavy (~100 KiB compressed) so in an ideal world I would use a more lightweight library like [shree](https://www.npmjs.com/package/shree) (which is more like 7 KiB compressed and has about the same functionality needed). The main functionality used is generating plane geometry, managing an orthographic camera, and compiling/binding uniforms to a GLSL shader material, which could all be implemented on top of shree. However, because the rendering of the background can be deferred without being that noticeable, fetching Three can be made low-priority and async, so the size of the library isn't a big deal.

Additionally, this PR includes a new page, `/fun/wave-playground/`, which uses a [Leva](https://github.com/pmndrs/leva)-powered panel that shows a set of controls to play around with the inputs to the canvas.

Finally, this PR includes a fork of `gatsby-plugin-dark-mode` that fixes various issues, [@skagami/gatsby-plugin-dark-mode](https://www.gatsbyjs.com/plugins/@skagami/gatsby-plugin-dark-mode/).